requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
starlette==0.38.5
pydantic==2.9.1
pydantic-settings==2.5.2
sqlalchemy==2.0.34
sqlalchemy_utils==0.41.2
black==24.8.0
ruff==0.6.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
uuid==1.30